# specs/1-todo-fullstack-app/plan/contracts/todo-api-openapi.yaml
openapi: 3.0.0
info:
  title: Todo API
  version: 1.0.0
  description: API for Todo application with user authentication and authorization

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.todoapp.com
    description: Production server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegistration:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          minLength: 8
          description: User's password
      required:
        - email
        - password

    UserLogin:
      type: object
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
      required:
        - email
        - password

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        created_at:
          type: string
          format: date-time
          description: When the user account was created
      required:
        - id
        - email
        - created_at

    TodoRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Title of the todo item
        description:
          type: string
          description: Detailed description of the todo item
        due_date:
          type: string
          format: date-time
          description: Optional deadline for the todo item
      required:
        - title

    TodoResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique todo identifier
        title:
          type: string
          description: Title of the todo item
        description:
          type: string
          description: Detailed description of the todo item
        completed:
          type: boolean
          default: false
          description: Whether the todo is completed
        due_date:
          type: string
          format: date-time
          description: Optional deadline for the todo item
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this todo
        created_at:
          type: string
          format: date-time
          description: When the todo was created
        updated_at:
          type: string
          format: date-time
          description: When the todo was last updated
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
          description: Title of the todo item
        description:
          type: string
          description: Detailed description of the todo item
        completed:
          type: boolean
          description: Whether the todo is completed
        due_date:
          type: string
          format: date-time
          description: Optional deadline for the todo item

    TodoToggleComplete:
      type: object
      properties:
        completed:
          type: boolean
          description: Whether the todo should be marked as completed

    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
      required:
        - detail

  parameters:
    todoId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique identifier of the todo item

  responses:
    UnauthorizedError:
      description: Unauthorized access - JWT token missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not authenticated"
    ForbiddenError:
      description: Access forbidden - user does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Access forbidden"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Todo not found"

paths:
  /api/auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with the provided email and password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '200':
          description: User successfully registered, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/login:
    post:
      summary: Authenticate user
      description: Authenticates user with email and password, returns JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Successfully authenticated, returns JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token for authentication
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/todos:
    get:
      summary: Get user's todos
      description: Retrieve all todos for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully retrieved todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create a new todo
      description: Creates a new todo for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoRequest'
      responses:
        '201':
          description: Todo successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /api/todos/{id}:
    get:
      summary: Get a specific todo
      description: Retrieve a specific todo by ID for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/todoId'
      responses:
        '200':
          description: Successfully retrieved todo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    put:
      summary: Update a todo
      description: Updates an existing todo for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/todoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      summary: Delete a todo
      description: Deletes an existing todo for the authenticated user
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/todoId'
      responses:
        '204':
          description: Todo successfully deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /api/todos/{id}/complete:
    patch:
      summary: Toggle todo completion status
      description: Updates the completion status of a todo item
      tags:
        - Todos
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/todoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoToggleComplete'
      responses:
        '200':
          description: Todo completion status successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'